requirements.txt
attrs==23.1.0
blinker==1.6.3
cattrs==23.1.2
click==8.1.7
coremltools==7.0
exceptiongroup==1.1.3
Flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
numpy==1.21.4
opencv-python==4.8.1
packaging==23.2
Pillow==10.1.0
protobuf==3.20.3
pyaml==23.9.7
PyYAML==6.0.1
sympy==1.12
tqdm==4.66.1
typing_extensions==4.8.0
Werkzeug==3.0.0
zipp==3.17.0
